memoize sidepanels for better rendering on theme change. Cleans up context and effects from RootLayout
